backend/requirements.txt
# FastAPI and WebSocket
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# ML / NLP
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
accelerate==0.25.0

# OpenAI
openai==1.10.0

# Data processing
pandas==2.1.4
numpy==1.26.3

# Environment
python-dotenv==1.0.0

# CORS
python-cors==1.0.0

# Async
aiofiles==23.2.1
